test(cli): integrate ACP console patcher cleanup tests into gemini_cleanup.test.tsx
